# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "email": "info@springonly.com",
      "name": "Equipo de Desarrollo de Spring Only"
    },
    "description": "API para información y gestión de créditos simples.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Credit API REST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://backend.springonly.com",
      "description": "Servidor Spring Only"
    }
  ],
  "components": {
    "schemas": {
      "CreateCustomerRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerTypeId": {
            "type": "string"
          },
          "riskCategoryId": {
            "type": "string"
          },
          "lineOfCreditAmount": {
            "type": "number"
          },
          "relationshipManagerId": {
            "type": "string"
          }
        }
      },
      "CreateCustomerResponse": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerTypeId": {
            "type": "string"
          },
          "riskCategoryId": {
            "type": "string"
          },
          "lineOfCreditAmount": {
            "type": "number"
          },
          "relationshipManagerId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "UpdateCustomerRequest": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string"
          },
          "customerTypeId": {
            "type": "string"
          },
          "riskCategoryId": {
            "type": "string"
          },
          "lineOfCreditAmount": {
            "type": "number"
          },
          "relationshipManagerId": {
            "type": "string"
          }
        }
      },
      "UpdateCustomerResponse": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerTypeId": {
            "type": "string"
          },
          "riskCategoryId": {
            "type": "string"
          },
          "lineOfCreditAmount": {
            "type": "number"
          },
          "relationshipManagerId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/customers": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Create customer",
        "description": "Create a new customer and return its created representation.",
        "operationId": "createCustomer",
        "requestBody": {
          "description": "Request payload for creating a new customer.",
          "content": {
            "application/json": {
              "examples": {
                "Create customer request": {
                  "summary": "A valid request to create a customer",
                  "value": {
                    "customerId": "CUST12345",
                    "customerName": "John Doe",
                    "customerTypeId": "MICRO EMPRESA",
                    "riskCategoryId": "NO",
                    "lineOfCreditAmount": 50000.0,
                    "relationshipManagerId": "RM001"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "Successful creation": {
                    "summary": "A successfully created customer",
                    "value": {
                      "customerId": "CUST12345",
                      "customerName": "John Doe",
                      "customerTypeId": "MICRO EMPRESA",
                      "riskCategoryId": "NO",
                      "lineOfCreditAmount": 50000.0,
                      "relationshipManagerId": "RM001",
                      "writtenAt": "2025-10-07T22:27:36-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CreateCustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid input": {
                    "summary": "An example of invalid input data",
                    "value": {
                      "message": "Invalid input data",
                      "code": "ERR400"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customer Resource"
        ]
      }
    },
    "/api/customers/{customerId}": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "summary": "Update customer",
        "description": "Update customer data identified by customerId.",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Request payload for updating a customer.",
          "content": {
            "application/json": {
              "examples": {
                "Update customer request": {
                  "summary": "A valid request to update a customer",
                  "value": {
                    "customerName": "John Doe Updated",
                    "customerTypeId": "MICRO EMPRESA",
                    "riskCategoryId": "NO",
                    "lineOfCreditAmount": 60000.0,
                    "relationshipManagerId": "RM001"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "examples": {
                  "Successful update": {
                    "summary": "A successfully updated customer",
                    "value": {
                      "customerId": "CUST12345",
                      "customerName": "John Doe Updated",
                      "customerTypeId": "MICRO EMPRESA",
                      "riskCategoryId": "PP",
                      "lineOfCreditAmount": 60000.0,
                      "relationshipManagerId": "RM001",
                      "writtenAt": "2025-10-07T22:27:36-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/UpdateCustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "Invalid input": {
                    "summary": "An example of invalid input data",
                    "value": {
                      "message": "Invalid input data",
                      "code": "ERR400"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "examples": {
                  "Customer not found": {
                    "summary": "An example of a non-existent customer",
                    "value": {
                      "message": "Customer not found",
                      "code": "ERR404"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customer Resource"
        ]
      },
      "get": {
        "summary": "Get customer by id",
        "description": "Retrieve customer details by customerId.",
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully",
            "content": {
              "application/json": {
                "examples": {
                  "Customer details": {
                    "summary": "Details of an existing customer",
                    "value": {
                      "customerId": "CUST12345",
                      "customerName": "John Doe",
                      "customerTypeId": "MICRO EMPRESA",
                      "riskCategoryId": "DU",
                      "lineOfCreditAmount": 50000.0,
                      "relationshipManagerId": "RM001",
                      "writtenAt": "2024-12-14T19:37:26-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/GetCustomerByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "examples": {
                  "Customer not found": {
                    "summary": "An example of a non-existent customer",
                    "value": {
                      "message": "Customer not found",
                      "code": "ERR404"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customer Resource"
        ]
      }
    },
    "/api/customers/{customerId}/loans": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "List loans by customer",
        "description": "List loans associated with the specified customerId.",
        "operationId": "listLoansByCustomerId",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loans list for customer",
            "content": {
              "application/json": {
                "examples": {
                  "Customer loans": {
                    "summary": "A list of loans for a customer",
                    "value": [
                      {
                        "loanId": 1,
                        "customerId": "CUST12345",
                        "currencyId": "PEN",
                        "principalAmount": 10000.0,
                        "interestRate": 5.5,
                        "loanDisbursementDate": "2023-01-01",
                        "numberOfMonthlyPayments": 12,
                        "loanStateId": "VIGENTE",
                        "writtenAt": "2025-09-28T11:15:00-05:00"
                      },
                      {
                        "loanId": 2,
                        "customerId": "CUST12345",
                        "currencyId": "USD",
                        "principalAmount": 20000.0,
                        "interestRate": 6.0,
                        "loanDisbursementDate": "2023-02-01",
                        "numberOfMonthlyPayments": 24,
                        "loanStateId": "DESAPROBADO",
                        "writtenAt": "2025-08-30T03:49:02-05:00"
                      }
                    ]
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetLoanByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No loans found for customer",
            "content": {
              "application/json": {
                "examples": {
                  "No loans found": {
                    "summary": "An example of a customer with no loans",
                    "value": {
                      "message": "No loans found for customer",
                      "code": "ERR404"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customer Resource"
        ]
      }
    },
...
    "/api/relationship-managers/{relationshipManagerId}/customers": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "List customers",
        "description": "List customers assigned to a specific relationship manager.",
        "operationId": "listCustomersByRelationshipManagerById",
        "parameters": [
          {
            "name": "relationshipManagerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customers list for relationship manager",
            "content": {
              "application/json": {
                "examples": {
                  "Customers list": {
                    "summary": "A list of customers for a relationship manager",
                    "value": [
                      {
                        "customerId": "CUST12345",
                        "customerName": "John Doe",
                        "customerTypeId": "MICRO EMPRESA",
                        "riskCategoryId": "DU",
                        "lineOfCreditAmount": 50000.0,
                        "relationshipManagerId": "RM001",
                        "writtenAt": "2024-12-14T19:37:26-05:00"
                      },
                      {
                        "customerId": "CUST67890",
                        "customerName": "María López",
                        "customerTypeId": "PERSONAL",
                        "riskCategoryId": "PP",
                        "lineOfCreditAmount": 15000.0,
                        "relationshipManagerId": "RM001",
                        "writtenAt": "2025-01-20T15:00:00-05:00"
                      }
                    ]
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetCustomerByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No customers found for relationship manager",
            "content": {
              "application/json": {
                "examples": {
                  "No customers found": {
                    "summary": "No customers for given relationship manager",
                    "value": {
                      "message": "El Ejecutivo de Cuentas no tiene clientes asignados",
                      "code": "RM002"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Relationship Manager Resource"
        ]
      }
    }
  }
}